xr::Asset::Manager and various asset types fixes and improvements.
- simplified scoping: failures in building and loading now return immediately;
- fixed integer usage (Texture creation and binding matches the Gfx API; the internal flags are uint16_t) / implicit casts;
- fixed multiple instances of name shadowing;
- SpriteSheet and TexturePack building no longer copy the sprite name to a buffer, rather than hashing the contents of the XML buffer;
- added xr::Shader::SetComponentInternal(), which OnLoaded() and SetComponent() has shared use of, enabling both to have their specific implementation;
- added handling of invalid bpp values in xr::Texture;
- removed names of unused parameters (usually 'dependencies');
- removed LTRACEIF()s;
- deprecated xr::TexturePack;

+++ xr3/h/xr/Texture.hpp
@@ -93,11 +93,11 @@ public:
   /// created with the given data.
   ///@note Must be called from the render thread.
   bool Upload(Gfx::TextureFormat format, uint16_t width, uint16_t height,
-    Gfx::FlagType createFlags, uint32_t numBuffers, Buffer const* buffers);
+    Gfx::FlagType createFlags, uint8_t numBuffers, Buffer const* buffers);
 
   ///@brief Binds texture to its target, for the given texture @a stage.
   ///@note Must be called from the render thread.
-  void Bind(uint32_t stage) const;
+  void Bind(uint8_t stage) const;
 
 private:
   // data
